pallet_price_aggregator::pallet

Trait Config

pub trait Config: Config + Config<RuntimeEvent: From<Event<Self>>> {
    type MaxValuesPerBlock: Get<u32>;
    type ProcessBlockValues: ProcessBlockValues;
    type NativeCurrencyId: Get<CurrencyId>;
    type CircularBufferLength: Get<u32>;
    type AggregationDuration: Get<BlockNumberFor<Self>>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MaxValuesPerBlock: Get<u32>

Maximum number of distinct currency values we can store during a single block.

type ProcessBlockValues: ProcessBlockValues

Used to process accumulated values in the current block.

type NativeCurrencyId: Get<CurrencyId>

Native currency ID that this pallet is supposed to track.

type CircularBufferLength: Get<u32>

Maximum length of the circular buffer used to calculate the moving average.

type AggregationDuration: Get<BlockNumberFor<Self>>

Duration of aggregation period expressed in the number of blocks. During this time, currency values are aggregated, and are then used to calculate the average value.

type WeightInfo: WeightInfo

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.